requirements.txt
streamlit==1.32.0
mysql-connector-python==8.0.32
pandas==2.0.0
python-dotenv==1.0.0
passlib==1.7.4